Fixes and improvements to the C++, Java DP libraries.
General:
- Add confidence intervals documentation
- Fix the GitHub landing REDME: we actually accept pull requests

C++ library:
- small refactoring

Java library:
- Fix budgeting of thresholdDelta for Count.computeThresholdedResult
- Add example image for some scenarios of the codelab.

GitOrigin-RevId: 88a2fcfcc641f30cf2b947fa2908340477711308

+++ java/main/com/google/privacy/differentialprivacy/BoundedSum.java
@@ -126,22 +126,26 @@ public class BoundedSum {
 
     resultReturned = true;
     double lInfSensitivity =
-            getLInfSensitivity(params.lower(), params.upper(), params.maxContributionsPerPartition());
+        getLInfSensitivity(params.lower(), params.upper(), params.maxContributionsPerPartition());
     noisedSum =
-            params
-                  .noise()
-                  .addNoise(sum, getL0Sensitivity(), lInfSensitivity, params.epsilon(), params.delta());
+        params
+            .noise()
+            .addNoise(sum, getL0Sensitivity(), lInfSensitivity, params.epsilon(), params.delta());
     return noisedSum;
   }
 
   /**
-   * ComputeConfidenceInterval computes a confidence interval that contains the true
-   * {@link BoundedSum} with a probability greater or equal to 1 - alpha using the noised
-   * {@link BoundedSum} computed by {@code computeResult()}.
+   * Computes a confidence interval that contains the true {@link
+   * BoundedSum} with a probability greater or equal to 1 - alpha using the noised {@link
+   * BoundedSum} computed by {@code computeResult()}.
+   *
+   * <p>See <a href="https://github.com/google/differential-privacy/tree/main/common_docs/confidence_intervals.md">
+   * the confidence intervals doc</a>.
    */
   public ConfidenceInterval computeConfidenceInterval(double alpha) {
     if (!resultReturned) {
-      throw new IllegalStateException("computeResult must be called before calling computeConfidenceInterval.");
+      throw new IllegalStateException(
+          "computeResult must be called before calling computeConfidenceInterval.");
     }
     ConfidenceInterval confInt =
         params
@@ -149,19 +153,16 @@ public class BoundedSum {
             .computeConfidenceInterval(
                 noisedSum,
                 params.maxPartitionsContributed(),
-                (double) params.maxContributionsPerPartition(),
+                params.maxContributionsPerPartition(),
                 params.epsilon(),
                 params.delta(),
                 alpha);
     if (params.lower() >= 0.0) {
-      confInt = ConfidenceInterval.create(
-              max(0.0, confInt.lowerBound()),
-              max(0.0, confInt.upperBound()));
-    }
-    else if (params.upper() <= 0.0) {
-      confInt = ConfidenceInterval.create(
-              min(0.0, confInt.lowerBound()),
-              min(0.0, confInt.upperBound()));
+      confInt =
+          ConfidenceInterval.create(max(0.0, confInt.lowerBound()), max(0.0, confInt.upperBound()));
+    } else if (params.upper() <= 0.0) {
+      confInt =
+          ConfidenceInterval.create(min(0.0, confInt.lowerBound()), min(0.0, confInt.upperBound()));
     }
     return confInt;
   }
